Make Payment View Parameter an ID
Previously, when viewing the payments of a specific user, the URL
pattern expected the username as a parameter. This caused problems for
users with non-alphanumerical characters in their username, preventing
the administration page from loading.

This commit fixes the regression, by altering the parameter of the
payments view to be an id instead of a username.