Bugfix for broken google login
Under certain circumstances, the 'Login with Google' button would not
work properly and stay grayed out. The cause of this was two fold, 1) a
race condition existed that dependended on the load order of certain
scripts and 2) an unsupported CSP directive on the Google side prevented
the google api in loading under Safari.

This commit fixes the issue by delaying the inclusion of the Google
Script until the page has loaded using window.onload.