atlas: show burthday != show age
Previously, when an alumni selected showing their birthday, the entire
date of birth including age was exposed to other users. This was not
clear enough in the setting, causing ages to be shown to other people
unexpectedly.

This commit updates the behaviour to only expose the day and month of
the date of birth instead of showing the complete date.